update/`applySearchFilter()` (virtual columns)
Extends the `applySearchFilter($query)`-function to check if the name of the `$whereColumn` exists on the (parent) schema. If the column (name) exists, add the table-name as a prefix `[tableName].[columnName]` just like before. If the column name does not exist, just take the column name `[columnName]`. If it still does not exists (as a **virtual column**), it will fail just like before.